# server/api/openapi.yaml
openapi: 3.1.0
info:
  title: Nuntius Server API
  version: 0.1.0
  description: |
    手机/平板远程控制 API。除明确标注的认证入口外，请求使用 HttpOnly
    `nuntius_session` Cookie；所有状态修改还需要 `X-CSRF-Token`。会产生设备
    副作用的请求同时必须带 `Idempotency-Key`，新命令仅在设备在线时接受；
    已接受命令在断线后由持久队列重放。
servers:
  - url: /api/v1
security:
  - webSession: []
paths:
...
  /client-releases/{releaseId}/{fileName}:
    get:
      security: []
      operationId: downloadClientRelease
      parameters:
        - {name: releaseId, in: path, required: true, schema: {type: string}}
        - {name: fileName, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: 不可变的 macOS ARM Client 更新包, content: {application/gzip: {schema: {type: string, format: binary}}}}
        '404': {description: 发布不存在}
...
components:
  securitySchemes:
    webSession: {type: apiKey, in: cookie, name: nuntius_session}
    deviceBearer: {type: http, scheme: bearer}
...